# Matching messages acceleration