--- a/whisper/whisperv6/filter.go
+++ b/whisper/whisperv6/filter.go
@@ -89,6 +89,7 @@ func (fs *Filters) Install(watcher *Filter) (string, error) {
 	}
 
 	fs.watchers[id] = watcher
+	//add topic matching for watcher
 	fs.topicMatcher.addFilterToTopicsMapping(watcher, id)
 	return id, err
 }
@@ -246,6 +247,7 @@ func IsPubKeyEqual(a, b *ecdsa.PublicKey) bool {
 	return a.X.Cmp(b.X) == 0 && a.Y.Cmp(b.Y) == 0
 }
 
+//topicMatcher keeps topic->watcher mapping
 type topicMatcher struct {
 	//structure - map[topic]map[filterID]
 	//mapping for topics
@@ -255,6 +257,7 @@ type topicMatcher struct {
 	pool   sync.Pool
 }
 
+//newTopicMatcher returns a newly created topic matcher
 func newTopicMatcher() *topicMatcher {
 	tm := new(topicMatcher)
 	tm.mapper = make(map[string]map[string]struct{})
@@ -265,9 +268,12 @@ func newTopicMatcher() *topicMatcher {
 	return tm
 }
 
+//take returns []string from pool
 func (fs *topicMatcher) take() []string {
 	return fs.pool.Get().([]string)
 }
+
+//resolve put []string to pool
 func (fs *topicMatcher) resolve(s []string) {
 	if cap(s) > MAX_POOL_CAPACITY {
 		return
@@ -275,6 +281,7 @@ func (fs *topicMatcher) resolve(s []string) {
 	fs.pool.Put(s[:0])
 }
 
+//addFilterToTopicsMapping fill topic->watcher mapping for current watcher
 func (fs *topicMatcher) addFilterToTopicsMapping(watcher *Filter, id string) {
 	fs.mx.Lock()
 	defer fs.mx.Unlock()
@@ -289,14 +296,16 @@ func (fs *topicMatcher) addFilterToTopicsMapping(watcher *Filter, id string) {
 	}
 }
 
-func (fs *topicMatcher) removeTopicFromTopicMapping(id string) {
+//removeTopicFromTopicMapping removes mapping info by filterID
+func (fs *topicMatcher) removeTopicFromTopicMapping(filterID string) {
 	fs.mx.Lock()
 	defer fs.mx.Unlock()
 	for i := range fs.mapper {
-		delete(fs.mapper[i], id)
+		delete(fs.mapper[i], filterID)
 	}
 }
 
+//prepareTopicsMapping returns set of topics for watcher
 func (fs *topicMatcher) prepareTopicsMapping(watcher *Filter) map[string]struct{} {
 	topics := make(map[string]struct{}, len(watcher.Topics))
 
@@ -312,6 +321,7 @@ func (fs *topicMatcher) prepareTopicsMapping(watcher *Filter) map[string]struct{
 	return topics
 }
 
+//matchedTopics write all matched topics to matched
 func (fs *topicMatcher) matchedTopics(topic TopicType, matched *[]string) {
 	fs.mx.RLock()
 	defer fs.mx.RUnlock()