NotifyWatchers has to iterate over all filters. It could become a bottleneck for users with a large number of filters.
Results of benchmark(8 topics per filter):

|   | 100 filters  | 50 filters  | 20 filters  | 10 filters  | 5 filters  |
|---|---|---|---|---|---|
| before  | 187707 ns/op  | 61981 ns/op  | 26762 ns/op  | 12554 ns/op  | 6283 ns/op  |
|  after | 424 ns/op  | 422 ns/op  |  412 ns/op | 406 ns/op  | 405 ns/op  |

Closes https://github.com/status-im/status-go/issues/642