@JekaMas 2018-02-27
@gluk256 When you have some time, please take a look at the PR. If all is OK, we're ready to implement the same fix for whisperV6.